+++ meta_package_manager/managers/apm.py
@@ -458,7 +458,12 @@ class APM(PackageManager):
         return outdated
 
     def upgrade_cli(self, package_id=None):
-        cmd = [self.cli_path, "update", "--no-confirm"]
-        if package_id:
-            cmd.append(package_id)
-        return cmd
+        return (
+            self.pre_cmds,
+            self.cli_path,
+            "update",
+            self.pre_args,
+            "--no-confirm",
+            self.post_args,
+            package_id,
+        )