+++ meta_package_manager/managers/opkg.py
@@ -189,7 +189,11 @@ class OPKG(PackageManager):
         return outdated
 
     def upgrade_cli(self, package_id=None):
-        cmd = [self.cli_path, "upgrade"]
-        if package_id:
-            cmd.append(package_id)
-        return cmd
+        return (
+            self.pre_cmds,
+            self.cli_path,
+            self.pre_args,
+            "upgrade",
+            self.post_args,
+            package_id,
+        )